change NREADA to BUFSIZA, always 8192
I was thinking about introducing reada.h into the API of zrec/dec.h.
It fits there well: the caller first needs to check the magic bytes,
and then dispatch the descriptor to the zrec backend.  On the other
hand, I still don't want to export the API on behalf of the shared
libraries that use it.  This is more of a convenience library (in
libtool terms).  So, basically, there can be a few reada.h header files
and a few reada.o instances buried in executables / shared objects, and
they all must interoperate.  This is doable, the buffer size becomes
part of the API, and cannot be easily changed.