return (size_t) -1 instead of ssize_t
This is a hard decision, but it solves the (conceptual) problem of
reading 2G+ data on 32-bit systems.  And it's not too bad in other
respects, only need to get used to it.  E.g. the condition

    (ret <= 0) // eof or error

now becomes

    (ret == 0 || ret == (size_t) -1)

The compiler can actually optimize the second condition to
something like

    (ret + 1 <= 1)

so it's still one branch.