stdio twist: sticky eof, 0 on error
It was fine to return (size_t) -1 on error and 0 on EOF, but what
to do with short reads?  When bytes read are fewer than requested,
there is no way to tell whether it is due to EOF or due to an error.
(Another option is to return -1 early, discarding the last chunk;
this makes things less general.  Yet another option is to utilize
the negative range and return -1 - total, but this makes 2G+ reads
impossible on 32-bit platforms.)

So I thought stdio's stance of returning 0 is not that bad after all.
But then you need to tell apart EOF and error, so it is only natural
to introduce those sticky flags.